requirements.txt
numpy
opencv
pillow
face-detection-tflite